requirements.txt
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.10.1
apache-airflow-providers-amazon==8.12.0
psycopg2-binary==2.9.9
boto3==1.34.9
python-dotenv==1.0.0
pandas==2.1.3
sqlalchemy>=1.4.28,<2.0
